+++ source/TestCommon/source/FunctionApp.TestCommon/ServiceBus/ListenerMock/AssertionException.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Runtime.Serialization;
-
 namespace Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ListenerMock;
 
 /// <summary>